# Fix type correction in HashJoin for nested low cardinality