--- a/tests/queries/0_stateless/02516_join_with_totals_and_subquery_bug.sql
+++ b/tests/queries/0_stateless/02516_join_with_totals_and_subquery_bug.sql
@@ -78,7 +78,7 @@ FROM (
     WITH TOTALS
 ) AS l
 FULL JOIN (
-    SELECT ([toString(number % 2)] :: Array(String)) AS item_id FROM numbers(3)
+    SELECT ([toString((number % 2) * 2)] :: Array(String)) AS item_id FROM numbers(3)
 ) AS r
 ON l.item_id = r.item_id
 ORDER BY 1,2,3
@@ -92,7 +92,7 @@ FROM (
     WITH TOTALS
 ) AS l
 FULL JOIN (
-    SELECT ([toString(number % 2)] :: Array(String)) AS item_id, count() FROM numbers(3) GROUP BY item_id
+    SELECT ([toString((number % 2) * 2)] :: Array(String)) AS item_id, count() FROM numbers(3) GROUP BY item_id
     WITH TOTALS
 ) AS r
 ON l.item_id = r.item_id
@@ -106,7 +106,7 @@ FROM (
     SELECT ([toString(number % 2)] :: Array(String)) AS item_id FROM numbers(3)
 ) AS l
 FULL JOIN (
-    SELECT ([toString(number % 2)] :: Array(LowCardinality(String))) AS item_id, count() FROM numbers(3) GROUP BY item_id
+    SELECT ([toString((number % 2) * 2)] :: Array(LowCardinality(String))) AS item_id, count() FROM numbers(3) GROUP BY item_id
     WITH TOTALS
 ) AS r
 ON l.item_id = r.item_id