… tests/queries/0_stateless/02516_join_with_totals_and_subquery_bug.reference