<!---
A technical comment, you are free to remove or leave it as it is when PR is created
The following categories are used in the next scripts, update them accordingly
utils/changelog/changelog.py
tests/ci/cancel_and_rerun_workflow_lambda/app.py
-->
### Changelog category (leave one):

- Bug Fix (user-visible misbehavior in an official stable release)


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
* Fix possible crash in joining on different types with LowCardinality inside and `WITH TOTALS`
* Slightly improve hash join performance (focusing a bit more on `LowCardinality` keys)

Close #57540 (I added minimized test case from that PR)