@vdimir 2023-12-12
This PR also slightly improves performance:



<details>
<summary>Setup</summary>

```
DROP TABLE IF EXISTS t1;
DROP TABLE IF EXISTS t2;

CREATE TABLE t1
(
    `id` Int64,
    `key1` String,
    `key2` LowCardinality(String)
) ENGINE = MergeTree() ORDER BY id;

INSERT INTO t1 SELECT number, toString( sipHash64(number, 1) % 1_000 ) as s, s FROM numbers(1_000_000);

CREATE TABLE t2
(
    `id` Int64,
    `key1` String,
    `key2` LowCardinality(String)
) ENGINE = MergeTree() ORDER BY id;

INSERT INTO t2 SELECT number, toString( sipHash64(number, 2) % 1_000 ) as s, s FROM numbers(1_000_000);
```
</details>

<details>
<summary>key1 String</summary>

```
-- current master
echo 'SELECT t1.id, t2.id FROM t1 JOIN t2 ON t1.key1 = t2.key1 FORMAT Null' | ~/clickhouse benchmark --port=9001 --iterations=10
localhost:9001, queries: 10, QPS: 0.262, RPS: 523683.766, MiB/s: 9.933, result RPS: 0.000, result MiB/s: 0.000.

0.000%          2.921 sec.
10.000%         2.945 sec.
20.000%         3.205 sec.
30.000%         3.269 sec.
40.000%         3.482 sec.
50.000%         4.192 sec.
60.000%         4.192 sec.
70.000%         4.210 sec.
80.000%         4.288 sec.
90.000%         4.530 sec.
95.000%         5.038 sec.
99.000%         5.038 sec.
99.900%         5.038 sec.
99.990%         5.038 sec.

-- patch
echo 'SELECT t1.id, t2.id FROM t1 JOIN t2 ON t1.key1 = t2.key1 FORMAT Null' | ~/clickhouse benchmark --port=9000 --iterations=10
localhost:9000, queries: 10, QPS: 0.280, RPS: 560158.893, MiB/s: 10.625, result RPS: 0.000, result MiB/s: 0.000.

0.000%          2.865 sec.
10.000%         2.866 sec.
20.000%         3.014 sec.
30.000%         3.178 sec.
40.000%         3.286 sec.
50.000%         3.422 sec.
60.000%         3.422 sec.
70.000%         3.587 sec.
80.000%         4.081 sec.
90.000%         4.366 sec.
95.000%         4.794 sec.
99.000%         4.794 sec.
99.900%         4.794 sec.
99.990%         4.794 sec.

```

</details>

<details>
<summary>key2 LowCardinality(String)</summary>

```
-- current master
echo 'SELECT t1.id, t2.id FROM t1 JOIN t2 ON t1.key2 = t2.key2 FORMAT Null' | ~/clickhouse benchmark --port=9001 --iterations=10
localhost:9001, queries: 10, QPS: 0.352, RPS: 703343.396, MiB/s: 6.708, result RPS: 0.000, result MiB/s: 0.000.

0.000%          2.461 sec.
10.000%         2.482 sec.
20.000%         2.540 sec.
30.000%         2.600 sec.
40.000%         2.691 sec.
50.000%         2.701 sec.
60.000%         2.701 sec.
70.000%         2.723 sec.
80.000%         2.945 sec.
90.000%         2.956 sec.
95.000%         4.034 sec.
99.000%         4.034 sec.
99.900%         4.034 sec.
99.990%         4.034 sec.

-- patch
echo 'SELECT t1.id, t2.id FROM t1 JOIN t2 ON t1.key2 = t2.key2 FORMAT Null' | ~/clickhouse benchmark --port=9000 --iterations=10
localhost:9000, queries: 10, QPS: 0.446, RPS: 891234.608, MiB/s: 8.499, result RPS: 0.000, result MiB/s: 0.000.

0.000%          1.807 sec.
10.000%         2.007 sec.
20.000%         2.186 sec.
30.000%         2.202 sec.
40.000%         2.254 sec.
50.000%         2.280 sec.
60.000%         2.280 sec.
70.000%         2.290 sec.
80.000%         2.326 sec.
90.000%         2.343 sec.
95.000%         2.533 sec.
99.000%         2.533 sec.
99.900%         2.533 sec.
99.990%         2.533 sec.

```
</details>